Move model class into data package
COSDK-441
