+++ core/src/main/java/com/adyen/checkout/core/data/model/PaymentMethodsApiResponse.kt
@@ -3,10 +3,10 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by ozgur on 30/4/2025.
+ * Created by ozgur on 1/5/2025.
  */
 
-package com.adyen.checkout.core.model
+package com.adyen.checkout.core.data.model
 
 import com.adyen.checkout.core.exception.ModelSerializationException
 import com.adyen.checkout.core.internal.data.model.ModelObject
@@ -28,7 +28,7 @@ data class PaymentMethodsApiResponse(
 ) : ModelObject() {
 
     companion object {
-//        private const val STORED_PAYMENT_METHODS = "storedPaymentMethods"
+        //        private const val STORED_PAYMENT_METHODS = "storedPaymentMethods"
         private const val PAYMENT_METHODS = "paymentMethods"
 
         @JvmField
@@ -40,7 +40,13 @@ data class PaymentMethodsApiResponse(
 //                            STORED_PAYMENT_METHODS,
 //                            serializeOptList(modelObject.storedPaymentMethods, StoredPaymentMethod.SERIALIZER)
 //                        )
-                        putOpt(PAYMENT_METHODS, serializeOptList(modelObject.paymentMethods, PaymentMethod.SERIALIZER))
+                        putOpt(
+                            PAYMENT_METHODS,
+                            serializeOptList(
+                                modelObject.paymentMethods,
+                                PaymentMethod.SERIALIZER,
+                            ),
+                        )
                     }
                 } catch (e: JSONException) {
                     throw ModelSerializationException(PaymentMethodsApiResponse::class.java, e)
