+++ core/src/main/java/com/adyen/checkout/core/sessions/CheckoutSessionInitializer.kt
@@ -10,8 +10,8 @@ package com.adyen.checkout.core.sessions
 
 import com.adyen.checkout.core.DispatcherProvider
 import com.adyen.checkout.core.Environment
+import com.adyen.checkout.core.data.model.Amount
 import com.adyen.checkout.core.internal.data.api.HttpClientFactory
-import com.adyen.checkout.core.model.Amount
 import com.adyen.checkout.core.sessions.data.api.SessionRepository
 import com.adyen.checkout.core.sessions.data.api.SessionService
 import kotlinx.coroutines.CoroutineDispatcher
